backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.29,<3
psycopg[binary]>=3.1.18
alembic>=1.13.1
pydantic>=2.6.4,<2.7
celery>=5.3.6
redis>=5.0.3
dateparser>=1.2.0
Babel>=2.14.0
langdetect>=1.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
weasyprint>=60.2
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
structlog>=24.1.0
sentry-sdk>=1.45.0
itsdangerous>=2.1.2
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiosqlite>=0.19.0
pydantic-settings>=2.0.3
email-validator>=2.1.1
